feat: create enzyme test and flatten props
