+++ packages/author-profile/__tests__/author-profile-content.native.test.js
@@ -1,0 +1,4 @@
+import test from "./author-profile-helper";
+import AuthorProfileContent from "../author-profile-content.js";
+
+test(AuthorProfileContent);
