+++ packages/author-profile/__tests__/author-profile-content.web.test.js
@@ -1,0 +1,4 @@
+import test from "./author-profile-helper";
+import AuthorProfileContent from "../author-profile-content.web.js";
+
+test(AuthorProfileContent);
